model/dr: [fix] disable traces by default
DynamoRIO can decide to merge code fragments into traces if it knows the
outcome of a jump instruction in advance. This however results in
instrumentation SEGFAULTs (namely, DynamoRIO ends up producing traces
that end with an unconditional branch, which is forbidden by the tool)
when exploring speculatve paths.

This patch disables traces formation (but not the code cache) to avoid
messing with the trace logic of DynamoRIO.